Dua folder len dau repo
