+++ NguyenVietHoang/junit/src/main/java/main/Student.java
@@ -1,0 +1,64 @@
+package main;
+
+import java.io.Serializable;
+
+public class Student implements Serializable {
+    private int id;
+    private String name;
+    private byte age;
+    private String address;
+    private double gpa;
+
+    public Student() {
+    }
+
+    public Student(int id, String name, byte age, 
+            String address, double gpa) {
+        super();
+        this.id = id;
+        this.name = name;
+        this.age = age;
+        this.address = address;
+        this.gpa = gpa;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public byte getAge() {
+        return age;
+    }
+
+    public void setAge(byte age) {
+        this.age = age;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public double getGpa() {
+        return gpa;
+    }
+
+    public void setGpa(double gpa) {
+        this.gpa = gpa;
+    }
+}
