--- Nhóm 15/NgoQuangManh/SoNguyenTo/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=4c0756ae
-build.xml.script.CRC32=8623e906
-build.xml.stylesheet.CRC32=8064a381@1.80.1.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=4c0756ae
-nbproject/build-impl.xml.script.CRC32=4ee6b4e4
-nbproject/build-impl.xml.stylesheet.CRC32=830a3534@1.80.1.48
