--- Nhóm 15/NgoQuangManh/SoNguyenTo/nbproject/private/private.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/E:/Code/Java/SoNguyenTo/test/Person/PersonTest.java</file>
-            <file>file:/E:/Code/Java/SoNguyenTo/src/Person/Person.java</file>
-        </group>
-    </open-files>
-</project-private>
