--- Nhóm 15/NguyenVietHoang/junit/src/main/java/main/SortStudentByGPA.java
@@ -1,13 +1,0 @@
-package main;
-
-import java.util.Comparator;
-
-public class SortStudentByGPA implements Comparator<Student> {
-
-    public int compare(Student student1, Student student2) {
-        if (student1.getGpa() > student2.getGpa()) {
-            return 1;
-        }
-        return -1;
-    }
-}
